Move "cheatcode" code out of `invites` module
